s3 source: Only mark objects seen after they have been successfully downloaded
Plausibly this is the fix for #7146, but I have not actually been able to
reproduce that issue following its instructions closely so I'm not certain.

This also introduces a new disabled testdrive test that fails without the patch
to the source but succeeds with it, it is an attempted minimization of the
various triggers described in the issue.