persist: relax BlobTraceBatch invariants to match differential.
Previously, we required that all updates in a trace batch have times between
[lower, upper). This invariant is overly restrictive with respect to compaction
policy, and in this commit we relax it to match differential's semantics.

Now, updates in a batch can have times in advance of upper if the since
frontier is in advance of upper. They cannot, however, have times in advance of
the since frontier.