Throw an error for unsupported outer joins in comma join (#7474)
Throw an error when the planner tries to push down the current
comma join to the leftmost side of a join tree involving either
right or full outer join, since otherwise the planner would build
a wrong join tree leading to wrong results.

Left joins are allowed since, even though the resulting plan may
not be optimal, it is semantically equivalent, ie. leads to the
same query results.