persist: refactor Future batch metadata
This commits extracts the key and description needed to reproduce a BlobFutureBatch
into its own struct, so that later work can add more metadata to that struct to
support efficient removals of old data from Future.

In addition, this commit adds a non-emptiness invariant for Future batch metadata,
which was present for BlobFutureBatch but not the corresponding metadata.