docs,kafka: remove requirement for infinite retention from docs
With the previous change that adds keys to BEGIN/END records in the
consistency topic, the consistency topic can have compaction enabled.
And we don't care about what happens to the data topic once we have
written out updates. It's completely up to the user what they do with
that data.