persist: evict batches with no useful data from Future
This commit teaches Future to track the maximum timestamp of updates for each batch
in that batch's metadata. This upper bound is used to determine when all data from a
batch has been absorbed into a Trace, at which time the future batch can be deleted.

This commit also gets rid of the contiguousness requirement for future batch sequence
number intervals. Maintaining that invariant would have added extra complexity to the
batch deletion code for extremely unclear benefits.