persist: remove Token
It's become clear that the initial model of a Token that can be
exchanged to construct a single differential dataflow operator is too
simplistic. The upcoming integration with coord will require the read
and write handles to be used in separate places.

Relatedly, add a Clone impl for write handles. The initial thinking was
that with a single registration per id and writer-exclusivity locks that
the user of write handle would be guaranteed it's the only writer to
that stream. However, the single registration requirement ended up being
awkward and unergonomic, so we removed it. As a result, there's no
remaining advantage to write handles not being Clone.