sql: simplify type conversion selection
When looking for valid casts, we return a "reflexive"/nop cast
when values shuld be considered equal. However, the code that
performed this check also became the place where we checked for
all situations where we should avoid doing work (e.g. when some
embedded values was None). This commit more stricly delineates
the responsibility for type equality to `get_cast`, and leaves
determining the work to do in each individual `CastTemplate`.