mzcompose: add code coverage feature
Add a --mz-coverage flag to mzcompose that will generate a code
coverage report for any contained Rust binaries. The flag works by
building a modified copy of the Rust binaries with the "-Z
instrument-coverage" flag, then wiring up the right environment
variables and volumes to emit the coverage report to the host machine
after running a workflow.