interchange: disallow Protobuf messages with unsigned integers
We are increasingly considering adding unsigned integer types to
Materialize. If we do so, we will certainly want to change the ingestion
of "uint32" and "uint64" types in Protobuf messages to their
corresponding unsigned types in Materialize.

We currently ingest these types as "numeric", which is capable of
representing all 32- and 64-bit unsigned integers, but loses information
about the "unsigned" constraint. This wart would be particularly
noticeable in a world with Protobuf sinks, where the following setup

    message Message {
      uint32 field f = 1;
    }

    CREATE SOURCE src ... FORMAT PROTOBUF MESSAGE '.Message' ...
    CREATE SINK ... FROM src FORMAT PROTOBUF MESSAGE '.Message' ...

would yield a sink that wrote Protobuf messages containing a decimal
type instead of uint32.

To leave our options open, this commit proposes disabling use of
Materialize with unsigned types entirely. While this might break some
users, we're almost certain we'll need to make a breaking change here
eventually, so better to do it sooner rather than later. If we do get
complaints, we can prioritize the unsigned integer work.