http-util: prepare for open-source release as mz-http-proxy
The http_proxy, https_proxy, and no_proxy support we built in the
http-util crate will be useful to at least one other person [0], and
probably many more. This commit prepares the crate for release to
crates.io.

Specifically:

  * The crate is renamed to "mz-http-proxy", for consistency with
    "mz-avro".

  * The license headers are updated to indicate the crate is Apache 2.0
    licensed.

  * The adapters are placed behind feature flags, so that you don't need
    to bring in reqwest if you just want hyper, or vice versa.

  * The documentation is improved, with a smidge of Materialize branding
    to boot.

I also switched `bin/doc` back to using the nightly toolchain in CI,
with support for optionally doing so locally too, to enable the
`doc_cfg` attributes that indicate which features must be enabled for
which items.

[0]: https://github.com/awslabs/aws-sdk-rust/issues/169#issuecomment-890276453