testdrive: Fix the parsing of "? EXPLAIN" actions
Previously, the ? EXPLAIN action would become confused by the
presence of a "!" on a Filter line, as "!" is considered
a special character in testdrive.

To avoid this, parse the ? EXPLAIN section of the .td file in
lines rather than characters and terminate only if the first
character of the line is a special character.

Add ? EXPLAIN to the documentation.