Rework per-worker CommandProcessedMetrics to reduce duplicate code
By combining enum_kinds and num_enum, we can change the worker command
representation to be an array of counters & values by which we increment
them. That reduces the number of names we have to duplicate in code, and
requires one fewer match statement on Command variants.

It also gets rid of some metrics that were represented but never got
incremented.