*: reduce all "deny" and "forbid" lints to "warn"
Lints at the deny/forbid level can be very frustrating when developing,
because you're forced to comply with the lint even when you're still
prototyping, just to get your code to compile.

This commit downgrades all these lints to "warn", to optimize for the
development experience. In CI, these lints are still enforced at the
"deny" level by our Clippy configuration, which upgrades all warnings to
hard errors.

The one glitch here is that `forbid` used to prevent submodules from
re-enabling the forbidden lint. I propose that we solve this socially,
by making it taboo to add `#![allow(missing_docs)]` unless you've gotten
buy-in from the overall owner of that area of the codebase.