coord: refactor catalog content migrations
Additionally, this refactor "undoes" the change committed in 36de7fca82. The
consequence of this is that any catalogs migrating from a version prior to 7.1
to anything past this commit will not have their confluent_wire_format set to
true; instead the options remains absent.

In the case that someone needs to remediate any breakage this causes, they
can proceed through the following binaries to get their catalog in the
expected state:

1. Pre 7.1 (initial state)
2. Some version including this commit, where they discover the error
3. 7.1
4. The current version