testdrive: remove support for corrupting avro/protobuf
Testdrive learned to write corrupt Avro and Protobuf data in #5995,
but... this feature is unnecessary. It is much easier to just write some
garbage bytes into the topics (that won't be valid Avro or Protobuf)
then go to the trouble of serializing correct Avro/Protobuf data and
then randomly permuting bytes.