persist: make storage locks reentrant for the "same" process after crash
To allow for restart without operator intervention in situations where
we've crashed and are unable to cleanly unlock (s3 has no flock
equivalent, for example), storage locks now support reentrance. We
define the "same" process as any that have the same reentrant_id, which
is an opaque user-provided string.

This, in essence, delegates the problem of ensuring writer-exclusivity
to the persist user, which may have access to better (possibly
distributed) locking primitives than we do.

Concretely, MZ Cloud will initially depend on the exclusivity of
attaching an EBS volume. We'll store the reentrant_id somewhere on the
EBS volume that holds the catalog. Any process that starts and has
access to this volume is guaranteed that the previous machine is no
longer available and thus the previous mz process is no longer running.
Similarly, a second process cannot accidentally start up pointed at the
same storage locations because it will not have the reentrant_id
available.

Violating writer-exclusivity will cause undefined behavior including
data loss. It's always correct, and MUCH safer, to provide a random
reentrant_id (e.g. a UUID). It will simply require operator intervention
to verify that the previous process is no longer running in the event of
a crash and to confirm this by manually removing the previous lock.