editorconfig: remove max_line_length directives
The max_line_length directive is not officially part of the EditorConfig
spec, and its current implementations have been contentious [0] [1]. The
tl;dr is that it is not clear whether to use max_line_length for soft
wrapping, hard wrapping, or just a visual display of a ruler, and
various tools and editors have different opinions.

Having the directive around is causing more trouble than its worth, in
my opinion. It overrode Vim's default line length for Git commit
messages [2], and caused isort to fight with black over the correct line
length [3]. There are probably other issues lurking.

So, this commit proposes removing the max_line_length directives from
our .editorconfig. We can reconsider if the EditorConfig project settles
on a meaning, or introduces a different directive like rulers [4].

[0]: https://github.com/editorconfig/editorconfig/issues/280
[1]: https://github.com/editorconfig/editorconfig/issues/387
[2]: https://github.com/MaterializeInc/materialize/commit/57574b6fcf1ae696f61a992458a27b7ef0c8e5eb
[3]: https://github.com/MaterializeInc/materialize/commit/5848bdd277f17d7fa20702dc90b8453a5d13cda9
[4]: https://github.com/editorconfig/editorconfig/issues/89