upgrade: Remove the shared volumes between different version workflows
With this change "upgrade from v0.7.3" will not have the catalog that was leftover
after the "test loading catalog from materialized-current" workflow ran at the
end of the "upgrade from v0.6.1" tests.

`docker-compose rm --volumes <container>*` only removes unnamed volumes, so it
was not doing anything about the existing `mzdata` directory.

The upgrade test framework was implicitly depending on every `check-from*` test
dropping items that cannot be parsed in any earlier version of materialized --
which are specifically the things that upgrade is most likely to test. Based on
the fact that the upgrade documentation doesn't state that you absolutely must
drop backwards-incompatible statements, I believe that this commit gets us
closer to the intended implementation. We could just update the docs, but a
genuine clean slate for each `upgrade-from` workflow seems easier to reason
about than a possibly populated one.