persist: fix incorrect loop in replay / read_to_end_flattened.
Previously, we were only sending data downstream for as long as the read call
indicated that there was more data to be read. Unfortunately, this breaks in the
case when data only lives in Buffer and the Buffer yields its data and indicates
that there is no more data to be read.

There's two separate things to fix here:

1. Future and Trace snapshots do not strictly follow the read function's requirement
   to only return true as long as there is additional data to be read, and instead
   return true as long as there is additional data to be read and one additional time.

2. The replay operator and read_to_end_flattened both need to change their reading
   strategy to disentangle the two steps of "polling read for more data" and
   "sending the data downstream". This change takes a naive approach for
   read_to_end_flattened which doubles the peak memory usage and a less naive
   approach for the replay operator.