persist: remove Buffer from Indexed
For now, send writes directly to Future. The immediate next step here will be to
execute sequences of writes, seals and allow compactions together and amortise
the costs of writing data / updating metadata on blob storage.

We are leaving Buffer all the way wired up in the Indexed object in the hopes that
we will bring it back soon.