misc/www: add sudo to oneliners
Unpacking a tarball into a nonwritable directory can fail with some very
obscure errors that have empirically resulted in confusion. Add `sudo`
to instructions to make it easier to copy/paste these commands.

One way to look at it is that people who run in configurations that
don't need sudo are likely to be savvy enough to remove sudo from the
commands, but the converse often isn't true.

Touches part of #7812.