mzbuild: take over all responsibility for managing RUSTFLAGS
mzbuild learned to manage the RUSTFLAGS environment variable when it
learned to inject flags to generate code coverage. Unfortunately, in
doing so, it was overwriting the important linker flags set in the
Docker image to use lld and to compress debug info.

Fix the bug by putting mzbuild entirely in charge of setting the correct
RUSTFLAGS.

Notably, this should make the code coverage build use a lot less memory,
and therefore actually complete without OOMing.