pgrepr,pgwire: entirely disable the list/map binary encodings
The binary encodings were previously a big hack that just used the text
encoding. We're comfortable not having binary encodings for these
custom types for now, but we don't want to commit ourselves to a *bad*
binary encoding, in case we want to introduce a proper binary encoding
in the future. So just prohibit using the binary encoding with these
types for now.

Fix #2997.
Fix #4628.