coord: correctly clean up read transactions on terminate
This affected clients that used terminate without explicitly ending their
transaction. In release mode, the telemetry reporter did this, which would
result in all of the system catalog objects from being compacted (due
to timedomains which brought in all the adjacent objects to mz_metrics).

I considered a fix that would move the antichain handles to the Session,
but using a common function to do transaction clearing seems like a 90%
solution to avoiding this kind of bug in the future, and with a much
smaller diff.

Fixes #7916