coord,pgwire: plumb down implicit transaction failures to pgwire
Previously we weren't sure where to send an error during an implicit
commit (at the end of a Query or Sync message). After direct comparison
with postgres, we have a clearer picture. I'm not sure how to exactly
induce this in postgres (have an INSERT succeed but then the implicit
transaction commit fail), but we can test a few nearby scenarios.

The difference we needed to observe is that for implicit transactions,
we don't care about the ExecuteResponse or tag at all, just that
an error happened. This greatly simplifies the error paths in
sequence_end_transaction and also means we don't need to change the tag
of the action should a failure occur.

Fixes #7926