persist: pipeline write, seal, and allow_compaction in nemesis
This starts to get nemesis coverage of command batching. This
essentially just means a bit of plumbing to move the `.recv()` call on
persist futures so that the nemesis Runner can pipeline them.

Most of the work is then fixing Validator's assumption that all steps
are executed serially. This surfaced mostly as it being deterministic
whether a given request should succeed or fail based on storage and
runtime availability. That worked when nemesis was totally
deterministic, but was already becoming a maintenance burden. As we move
toward pipelining requests, running them concurrently in threads, and
adding periods of non-determinism to the unreliable storage, it will
require trinary logic (yes/no/maybe), which gets surprisingly
complicated.

Stepping back, nemesis's purpose is to verify that we adhere to our
external API guarantees. This means assertions on histories of
request/response pairs and outputs of dataflow operators, not anything
around what to expect given detailed history of the availability of
external systems.

However, we do keep this one "liveness" check around so we don't
accidentally pass validation by simply erroring every request (which is
a valid, but not useful history).

Still TODO for the future:
- Concurrently driving requests from multiple threads.
- Making the rest of the request types asynchronous.