interchange,testdrive: unfork serde-protobuf
We forked serde-protobuf library to add one method to improve an error
message.  Empirically we have not been keeping up with maintenance of
our fork, so it's time to ditch the fork.

This commit swaps our fork for the latest official serde-protobuf
release. I left a TODO to restore the error message improvement if the
patch [0] from our fork is accepted upstream.

Notably, this commit removes the last usage of failure from our
codebase.

[0]: https://github.com/dflemstr/serde-protobuf/pull/9