persist: only increment compaction levels if the compacted batch is larger than parents
The trace maintenance code assigns a compaction level to each batch in the trace. The
compaction level indicates "this batch is the result of compacting 2^{level} batches"
and two consecutive batches (batches covering a consecutive time interval) at the
same compaction level L can be merged together to form a single batch at level L + 1.

This strategy helps ensure that we keep a logarithmic number of batches relative
to the total number of batches inserted into the trace, and that we perform a linear
number of compactions relative to the number of batches inserted in the trace.

This strategy works well in the case where data is mostly inserts, but breaks down
when retractions are common. Specifically, in the mz_metrics case, where all data
are retracted after a specified retention interval (default 300 seconds) we have
no reason to keep around data longer than 2x retention interval. However, the
However, the default policy of always promoting the merge of 2 batches at level L
to L + 1 means that e.g. two batches each representing an hours worth of time at
level 10 would create another batch representing two hours worth of time, at level
11, and would need to wait for approximately 2^11 more batches to come in before
it could be compacted further, even though all the retractions for all of the
data in the batch are most likely already present in the trace.

This fix here is simple -- only promote batches that are larger than both of their
parents. The fix gets to the main reason why we wanted to amortize the cost of future
compaction -- we were worried about doing a quadratic amount of work for large batches.
But if the batch isn't actually large, this isn't a concern and there's no future cost
to amortize for.