persist: ignore seal notifications before the snapshot in listen_source
We were previously crashing on startup due to an invalid attempt to downgrade
a capability at some time T to a time less than T (often less by 50 - 100 ms).
This happened because in the persisted source operator we:
- register a listener
- read a snapshot
- downgrade the capability to the snapshot's sealed frontier
- listen for updates

The last step now becomes a problem because we registered the listener before
taking the snapshot so theres a potential for some seal / data notifications to
come in even though they are not in advance of the snapshot's sealed frontier.
We also can't register the listener after taking the snapshot because there's
no guarantee that there won't be arbitrary delay between those two operations
which would cause us to lose data.

Instead, we simply need to check all seal notifications if they are in advance
of the snapshot's sealed frontier, and only apply them if so. We already do this
correctly for the data notifications

Touches #8015