persist: add persistent_upsert() operator
This is a generic version of upsert_core() in upsert.rs (which we
factored out before). It is generic because we want to keep the persist
crate free from dataflow/mz specifics, such as Row.

Contrary to upsert_core(), this operator doesn't support applying
filter/project operations right before we insert into our internal
upsert state. We don't support it, because this would require storing
Result<Row, DataflowError> in the backing persistent stream, and we
don't want to add a Codec for DataflowError for now.

We first create the operator like this and then later add the
persistence code so that it's easy to see what the actual changes needed
for persistence are.