expand kafka regression coverage
Adding regression coverage for confluent 5 (kafka 2) and confluent 4
(kafka 1).

The only thing that needs to be dropped for this is the zstd compression
test - I've taken the somewhat hacky but expedient step of moving that
to a separate td file.

I'm dropping coverage for the other 5.x versions. They seem redundant
and it'll free up capacity for these new checks, but I don't have strong
feelings one way or the other.