reduce metric interval parameters for kafka matrix test
We have tests in kafka-stats.td that check that these metrics are
populated. By default materialize has a fairly long collection interval,
causing these tests to fail. This brings the Kafka matrix test in line
with the interval config parameters we use for the testdrive suite.

Fixes nightly CI runs.