dataflow/logging: Rework Timely logging to consolidate early
This change fundamentally changes the implementation of the Timely logging
infrastructure. Instead of relying on Differential to retract accumulations
for past operators and dataflows, we maintain the information manually in
the demux operator and emit the corresponding retractions once a dataflow
or operator goes out of scope.

The benefit of this approach is that less data needs to be copied around
and it enables early consolidation of worker-local data within the demux
operator. The downside it that the implementation is less elegant and
requires potentially more care when maintaining.